Change Key Add Role

+++ VinaC2C.Data/Models/FeatureRole.cs
@@ -6,9 +6,9 @@ namespace VinaC2C.Data.Models
 {
     public class FeatureRole : ObjectBaseModel
     {
-        public int FeatureID { get; set; }
+        public Int64 FeatureID { get; set; }
 
-        public int UserID { get; set; }
+        public Int64 UserID { get; set; }
 
         public bool IsAllow { get; set; }
     }
